# docs/swagger.yaml
basePath: /
definitions:
  config.APIDBPoolConfig:
    properties:
      connMaxIdleTime:
        type: integer
      connMaxLifetime:
        type: integer
      default:
        type: boolean
      maxIdleConns:
        type: integer
      maxOpenConns:
        type: integer
      reconnect:
        type: boolean
    type: object
  config.EnvVars:
    type: object
  models.APIError:
    properties:
      hasError:
        type: boolean
      id:
        type: integer
      messages:
        items:
          $ref: '#/definitions/models.APIMessage'
        type: array
      msg:
        type: string
      result: {}
    type: object
  models.APIMessage:
    properties:
      code:
        type: string
      context:
        type: string
      description:
        type: string
      field:
        type: string
      messageType:
        type: integer
    type: object
  models.APIResponse:
    properties:
      codigo:
        type: string
      hasError:
        type: boolean
      id:
        type: integer
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      msg:
        type: string
    type: object
  models.CancelarCobrancaFrontendInput:
    properties:
      codigoLiquidacao:
        type: string
      id_convenio:
        type: integer
      id_proposta:
        type: integer
      id_proposta_parcela:
        type: integer
      id_securitizadora:
        type: integer
      numero_acompanhamento:
        type: string
      numero_ccb:
        type: integer
      url_webhook:
        type: string
    required:
    - id_proposta
    - id_proposta_parcela
    - numero_acompanhamento
    - numero_ccb
    - url_webhook
    type: object
  models.ConsultaBoleto:
    properties:
      descricaoBoletoRegistro:
        type: string
      dtCredito:
        type: string
      dtExpiracao:
        type: string
      linhaDigitavel:
        type: string
      liquidacao:
        type: boolean
      numeroBoleto:
        type: integer
      operacaoBoletoRegistro:
        type: integer
      qrCode:
        $ref: '#/definitions/models.ConsultaBoletoQrCode'
      situacaoBoletoRegistro:
        type: integer
      urlImpressao:
        type: string
    type: object
  models.ConsultaBoletoQrCode:
    properties:
      emv:
        type: string
      imagem:
        type: string
    type: object
  models.ConsultaCobrancaFrontEndInput:
    properties:
      id_convenio:
        type: integer
      id_proposta:
        type: integer
      id_securitizadora:
        type: integer
      numero_acompanhamento:
        type: string
      numero_parcela:
        type: integer
    required:
    - id_convenio
    - id_proposta
    - id_securitizadora
    - numero_acompanhamento
    type: object
  models.ConsultaCobrancaLancamento:
    properties:
      descLancamento:
        type: string
      dtLancamento:
        type: string
      dtVencimento:
        type: string
      informarFundo:
        type: boolean
      viaBoleto:
        type: boolean
      vlrAbatimento:
        type: integer
      vlrDesconto:
        type: integer
      vlrIOF:
        type: integer
      vlrJuros:
        type: integer
      vlrMora:
        type: integer
      vlrMulta:
        type: integer
      vlrPagamento:
        type: integer
      vlrParcela:
        type: integer
      vlrParcelaAtualizado:
        type: integer
      vlrSaldoParcela:
        type: integer
      vlrTarifas:
        type: integer
    type: object
  models.ConsultaCobrancaMessage:
    properties:
      code:
        type: string
      context:
        type: string
      description:
        type: string
      field:
        type: string
      messageType:
        type: integer
    type: object
  models.ConsultaCobrancaParcela:
    properties:
      boletos:
        items:
          $ref: '#/definitions/models.ConsultaBoleto'
        type: array
      codigoParcela:
        type: string
      dtVencimento:
        type: string
      dtVencimentoAtual:
        type: string
      lancamentos:
        items:
          $ref: '#/definitions/models.ConsultaCobrancaLancamento'
        type: array
      nroOrdemPagamento:
        type: string
      nroParcela:
        type: integer
      pix:
        items:
          $ref: '#/definitions/models.ConsultaPix'
        type: array
      situacao:
        type: integer
      vlrDesconto:
        type: number
      vlrParcela:
        type: number
      vlrSaldoAtual:
        type: number
      vlrTotalPago:
        type: number
    type: object
  models.ConsultaCobrancaResponse:
    properties:
      descricaoSituacaoAgenda:
        type: string
      descricaoTipoContrato:
        type: string
      documentoFederalCliente:
        type: string
      dtCancelamento:
        type: string
      dtInclusao:
        type: string
      dtLiquidacao:
        type: string
      ecCliente:
        type: string
      hasError:
        type: boolean
      messages:
        items:
          $ref: '#/definitions/models.ConsultaCobrancaMessage'
        type: array
      msg:
        type: string
      nomeCliente:
        type: string
      nomeEmpresa:
        type: string
      numeroProposta:
        type: integer
      parcelas:
        items:
          $ref: '#/definitions/models.ConsultaCobrancaParcela'
        type: array
      qtdeParcelas:
        type: integer
      result: {}
      situacaoAgenda:
        type: integer
      situacaoProposta:
        type: integer
      vlrFinanciado:
        type: number
      vlrRetencaoDiarioAprox:
        type: number
      vlrTotalPago:
        type: number
    type: object
  models.ConsultaPix:
    properties:
      dtCredito:
        type: string
      dtExpiracao:
        type: string
      dtVencimento:
        type: string
      emv:
        type: string
      imagem:
        type: string
      liquidacao:
        type: boolean
    type: object
  models.GerarCobrancaFrontendInput:
    properties:
      data_expiracao:
        type: string
      data_vencimento:
        type: string
      id_convenio:
        type: integer
      id_forma_cobranca:
        type: integer
      id_proposta:
        type: integer
      id_proposta_parcela:
        type: integer
      id_securitizadora:
        type: integer
      numero_acompanhamento:
        type: string
      numero_ccb:
        type: integer
      numero_parcela:
        type: integer
      token:
        type: string
      url_webhook:
        type: string
      valor_parcela:
        type: number
    required:
    - data_expiracao
    - data_vencimento
    - id_forma_cobranca
    - id_proposta
    - id_proposta_parcela
    - numero_acompanhamento
    - numero_ccb
    - numero_parcela
    - token
    - url_webhook
    type: object
  models.Message:
    properties:
      code:
        type: string
      context:
        type: string
      description:
        type: string
      field:
        type: string
      messageType:
        type: integer
    type: object
host: bmp-fgts.Confiapay.com.br
info:
  contact:
    email: felipeaug92@gmail.com
    name: Felipe Augusto De Faria Freitas
  description: Microserviço de acesso aos serviços de geração de cobranças do BMP.
    \n Para mais informações referentes aos dados a serem enviados, consulte a aba
    "model" na seção de body dos endpoints.
  title: BMP Cobrança
  version: "1.0"
paths:
  /cancelamento:
    post:
      consumes:
      - application/json
      description: Realiza o cancelamento de boletos.
      parameters:
      - description: Dados da cobrança.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CancelarCobrancaFrontendInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Cancelar Boleto.
      tags:
      - CreditoPessoal
  /config/db/set:
    post:
      consumes:
      - application/json
      description: Reconfigura os valores do DB.
      parameters:
      - description: Parâmetros do banco de dados.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/config.APIDBPoolConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Reconfigurar DB.
      tags:
      - Config
  /config/env:
    post:
      consumes:
      - application/json
      description: Configura variáveis de ambiente.
      parameters:
      - description: Variáveis que se deseja alterar.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/config.EnvVars'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Configurar variáveis de ambiente globais.
      tags:
      - Config
  /consulta:
    post:
      consumes:
      - application/json
      description: Realiza a consulta de cobranças geradas para uma proposta específica.
      parameters:
      - description: Identificador da proposta.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ConsultaCobrancaFrontEndInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConsultaCobrancaResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Consultar Cobranças.
      tags:
      - CreditoPessoal
  /geracao:
    post:
      consumes:
      - application/json
      description: Realiza a geracao de cobranças na API do BMP.
      parameters:
      - description: Dados da cobrança.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.GerarCobrancaFrontendInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.APIError'
      summary: GerarCobranca Proposta.
      tags:
      - CreditoPessoal
swagger: "2.0"